+++ source/library/cssrule-enumerations.js
@@ -1,0 +1,19 @@
+// CSSRule type enums
+export default {
+	unknown: 0,
+	style: 1,
+	charset: 2,
+	import: 3,
+	media: 4,
+	fontface: 5,
+	page: 6,
+	keyframes: 7,
+	keyframe: 8,
+	namespace: 9,
+	counter: 11,
+	supports: 12,
+	document: 13,
+	fontfeature: 14,
+	viewport: 15,
+	region: 16
+};
