+++ source/library/get-keyframes.js
@@ -1,137 +1,5 @@
-// CSSRule type enums
-const enums = {
-	unknown: 0,
-	style: 1,
-	charset: 2,
-	import: 3,
-	media: 4,
-	fontface: 5,
-	page: 6,
-	keyframes: 7,
-	keyframe: 8,
-	namespace: 9,
-	counter: 11,
-	supports: 12,
-	document: 13,
-	fontfeature: 14,
-	viewport: 15,
-	region: 16
-};
-
-const empty = [];
-
-/**
- * Cast array-like objects and collections to Array
- * @param  {Object} arrayLike array-like to cast to Array
- * @return {Array} Array cast from arrayLike
- * @private
- */
-function toArray(arrayLike) {
-	return empty.slice.call(arrayLike); // eslint-disable-line prefer-reflect
-}
-
-/**
- * Gets all CSSRules of type typeName and matching the predecate from rules
- * @param  {string} [typeName='style']    - CSSRule type to search for, valid types:
- *                                          unknown, style, charset, import, media, fontface, page, keyframes, keyframe, namespace, counter, supports, document, fontfeature, viewport, region
- * @param  {array} [rules=[]]             - Array of CSSRules to search
- * @param  {function} [predecate=Boolean] - Predecate function to filter matches
- * @return {array} Array of matching CSSRules
- * @private
- */
-export function getDeclarations(typeName = 'style', rules = [], predecate = Boolean) {
-	// Get target type enum
-	const type = enums[typeName];
-
-	return toArray(rules)
-		// filter by rule type
-		.filter(rule => rule.type === type)
-		// filter with user-provided predecate
-		.filter(predecate)
-		// unwrap cssRules
-		.map(rule => rule.cssRules)
-		// flatten cssRules
-		.reduce((results, cssRules) => {
-			return [
-				...results,
-				...toArray(cssRules)
-			];
-		}, []);
-}
-
-/**
- * Gets all KeyFrameRule declarations attached to CSS animationName preset in rules
- * @param  {string} animationName - CSS animationName to search KeyFrameRule declarations for
- * @param  {array} rules - Array of CSSRules to search
- * @return {array} Array of matching KeyFrameRules
- * @private
- */
-export function getKeyframeDeclarations(animationName, rules) {
-	// Filter for KeyFrameRules matching an animationName
-	return getDeclarations('keyframes', rules, rule => rule.name === animationName);
-}
-
-/**
- * Parses KeyFrameRule.keyText to an array of integers holding keyframe percentages
- * @param  {string} keyText KeyFrameRule.keyText to parse
- * @return {array}          Array of percentages for this KeyFrameRule
- * @private
- */
-export function parseKeyframeKey(keyText) {
-	// Split multivalue key,
-	return keyText.split(',')
-		// Trim any remaining whitespace
-		.map(key => key.trim())
-		// "Understand" CSS keyText keywords
-		.map(key => key
-			.replace('from', '0')
-			.replace('to', '100'))
-		// Remove any math symbols
-		.map(key => key.replace('%', ''))
-		// Parse to integer
-		.map(key => parseInt(key, 10));
-}
-
-/**
- * Gets map of defined styles from CSS2Properties object
- * @param  {CSS2Properties} properties CSS2Properties object to return defined styles from
- * @return {object}       plain object containing defined styles as key value pairs
- * @private
- */
-export function getDefinedStyles(properties) {
-	const styles = {};
-
-	for (let i = properties.length - 1; i >= 0; i -= 1) {
-		const name = properties.item(i);
-		const value = properties.getPropertyValue(name);
-		if (value !== 'initial') {
-			styles[name] = value;
-		}
-	}
-
-	return styles;
-}
-
-/**
- * Transforms KeyFrameRule to array of web animation compatible keyframes
- * @param  {Object} keyFrameRule KeyFrameRule to transform
- * @return {Array}               Array of webanimation keyframes
- * @private
- */
-export function transformKeyframeDeclaration(keyFrameRule) {
-	// Convert keyFrame.keyText to integers holding percentage of keyframe
-	const percentages = parseKeyframeKey(keyFrameRule.keyText);
-	const style = getDefinedStyles(keyFrameRule.style);
-
-	return percentages.map(percentage => {
-		return {
-			// Convert percentage to fraction of 1 for webanimation compat
-			offset: percentage / 100,
-			// Mixin with extracted keyframe styling
-			...style
-		};
-	});
-}
+import getKeyframeDeclarations from './get-keyframe-declarations';
+import transformKeyframeDeclaration from './transform-keyframe-declaration';
 
 /**
  * Gets webanimation keyframes attached to a CSS animationName
