+++ source/test/unit/get-defined-styles.js
@@ -3,7 +3,7 @@ import tape from 'tape';
 import simpleAnimation from './fixtures/simple-animation-declaration.js';
 const simpleAnimationKeyFrameRules = simpleAnimation[0].cssRules[0].cssRules;
 
-import {getDefinedStyles} from '../../library/get-keyframes.js';
+import getDefinedStyles from '../../library/get-defined-styles.js';
 
 tape('get-defined-styles', t => {
 	t.ok(
